Disconnect and free session's resources on __del__
The use may call session.connect() and forget to call disconnect(). It
is expected in Python that all the resources be released when the
session object is garbage-collected.

Call disconnect() on __del__().

Because __del__() may be called when the Python VM is shutting down,
there are no guarantees about the validity of global objects, specially
modules. Therefore, __del__() cannot relay on the 'api' module so an
explicit reference to api.Api is kept during the lifetime of the session
object, used on __del__() and released later.